Move MDevType to mdev module
